ft_printf complete, before refactoring

+++ printf/include/libft.h
@@ -6,7 +6,7 @@
 /*   By: junhseo <junhseo@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 14:54:26 by junhseo           #+#    #+#             */
-/*   Updated: 2023/04/07 15:18:33 by junhseo          ###   ########.fr       */
+/*   Updated: 2023/04/10 13:18:30 by junhseo          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@ int		ft_strlen(const char *s);
 char	*ft_strchr(const char *s, int c);
 void	put_chr(char c, int *count);
 void	put_str(char *s, int *count);
-void	put_nbr(int n, int *count);
+void	put_nbr(long n, int *count);
 
 #endif
